test: use config duration type in newborn-sex regression

+++ crates/anthrosim-core/tests/newborn_sex_label_invariance.rs
@@ -123,7 +123,7 @@ fn sex_key(sex: ReproductiveSex) -> usize {
 
 fn run(
     seed: u64,
-    years: u32,
+    years: u64,
     width: u32,
     initial_population: u32,
     founders: FounderPopulationDefinition,
@@ -185,7 +185,7 @@ fn run(
 
 fn run_two_households(
     seed: u64,
-    years: u32,
+    years: u64,
     swapped_labels: bool,
     male_birth_permille: u16,
 ) -> PhysicalOutcome {
